refactor: init moving server to bot
