+++ packages/bot/src/entities/help-channel.ts
